Driver code & mock implementations of ProblemLoader and SubmissionRunner
All of the code should be considered preliminary, but the only logic that is *explicitly* a mock implementation is ProblemLoader and SubmissionRunner.

The example.py script should execute using the mock driver returned by get_mock_driver() which uses the base implementations of ProblemLoader and SubmissionRunner.

New subclasses should be created for ProblemLoader and SubmissionRunner to be concrete implementations.

The LLM2PP_Driver should *also* be considered a base class, but only because it directly accepts ProblemLoader and SubmissionRunner types. Ideally, we will have a separate subclass of LLM4PP_Driver for each evaluation pipeline which constructs the proper ProblemLoader and SubmissionRunner for that specific pipeline.